[WIP] Changed the fitness function and a lot more

+++ WRCIComponentChooser/Program.cs
@@ -12,18 +12,18 @@ namespace WRCIComponentChooser
     {
         static void Main(string[] args)
         {
-            GA ga = new GA(6, 1000, Fitness);
+            GA ga = new GA(3, 500, Fitness);
             ga.Train(100);
             Console.WriteLine(ga.BestIndividual.Fitness);
         }
 
 
-        static double Fitness(double[] values, int pid)
+        static double Fitness(double[] values, int pid, bool delete)
         {
-            if (values.Length != 6)
+            if (values.Length != 3)
                 throw new ArgumentException("Should have six component values");
 
-            string[] names = new string[] { "r1v", "r2v", "r3v", "r4v", "c1v", "c2v" };
+            string[] names = new string[] { "r1v", "r2v", "cv" };
 
             using (var sw = new StreamWriter("multivibrator." + pid + ".net"))
             {
@@ -33,7 +33,7 @@ namespace WRCIComponentChooser
                     while (!sr.EndOfStream)
                     {
                         string l = sr.ReadLine();
-                        for (int i = 0; i < 6; i++)
+                        for (int i = 0; i < 3; i++)
                         {
                             if (l.Contains(names[i]))
                             {
@@ -71,69 +71,31 @@ namespace WRCIComponentChooser
                 }
             }
 
-            const double desiredFrequency = 2;
-            const double desiredDutyCycle = 0.5;
+            const decimal desiredPeriod = 5M;
+            const decimal desiredDutyCycle = 0.5M;
 
             double error = 0;
-
-            foreach (Vector2 p in simValues)
+            using (StreamWriter w = new StreamWriter("wave." + pid + ".csv"))
             {
-                double expected = 5 * ((p.T % desiredFrequency < desiredDutyCycle * desiredFrequency) ? 1 : 0);
-                error += Math.Abs(expected - p.V);
+                foreach (Vector2 p in simValues)
+                {
+                    double expected = 5 * (((decimal)p.T % desiredPeriod < desiredDutyCycle * desiredPeriod) ? 1 : 0);
+                    w.WriteLine("{0}\t{1}\t{2}", p.T, expected, p.V);
+                    error += Math.Abs(expected - p.V);
+                }
             }
-
-            /*
-            //calculate duty cycle and frequency
-            int zeroCrossings = 0;
-            double timeUnderZero = 0;
-            double timeOverZero = 0;
-            double average = simValues.Average(v => v.V);
-
-            for (int i = 0; i + 1 < simValues.Count; i++)
-            {
-                if (simValues[i].V <= average && simValues[i + 1].V >= average)
-                {
-                    //TODO: interpolate a crossing point and add to time under/over zero
-                    zeroCrossings++;
-                }
-                else if (simValues[i].V >= average && simValues[i + 1].V <= average)
-                {
-                    //TODO: interpolate a crossing point and add to time under/over zero
-                    zeroCrossings++;
-                }
-                else if (simValues[i].V > average && simValues[i].V > average)
-                {
-                    timeOverZero += (simValues[i + 1].T - simValues[i].T);
-                }
-                else if (simValues[i].V < average && simValues[i].V < average)
-                {
-                    timeUnderZero += (simValues[i + 1].T - simValues[i].T);
-                }
-                else
-                {
-                    throw new Exception("My logic must be wrong");
-                }
-
-            }
-
-
-            double dutyCycle = timeOverZero / (timeOverZero + timeUnderZero);
-            double frequency = (zeroCrossings / (simValues.Last().T - simValues.First().T)) / 2;
-
-            
-            
-            double fitness = frequency; //Math.Abs(desiredFrequency - frequency) / desiredFrequency;// +Math.Abs(desiredDutyCycle - dutyCycle); //Math.Sqrt(Math.Pow(desiredFrequency - frequency, 2) + Math.Pow(desiredDutyCycle - dutyCycle, 2));
-            */
 
             double fitness = -error / simValues.Count;
 
-            Console.WriteLine("Frequency: {0:N3}\tDuty Cycle: {1:N4}\tFitness: {2:N4}", 0, 0, fitness);
-
-            File.Delete("multivibrator." + pid + ".net");
-            File.Delete("multivibrator." + pid + ".log");
-            File.Delete("multivibrator." + pid + ".raw");
-            File.Delete("multivibrator." + pid + ".op.raw");
-
+            Console.WriteLine("{0}\tFitness: {1:N4}", pid, fitness);
+            if (delete)
+            {
+                File.Delete("multivibrator." + pid + ".net");
+                File.Delete("multivibrator." + pid + ".log");
+                File.Delete("multivibrator." + pid + ".raw");
+                File.Delete("multivibrator." + pid + ".op.raw");
+                File.Delete("wave." + pid + ".csv");
+            }
             return fitness;
         }
     }
